requirements.txt
beautifulsoup4==4.13.4
chromadb==1.5.7
lxml==5.4.0
llama-index-core==0.14.20
llama-index-embeddings-huggingface==0.7.0
llama-index-llms-huggingface==0.7.0
llama-index-llms-ollama==0.10.1
llama-index-postprocessor-flag-embedding-reranker>=0.5.0,<0.6
llama-index-vector-stores-chroma==0.5.5
pandas==2.3.3
sec-edgar-downloader==5.1.0
sentence-transformers==5.4.0
streamlit==1.56.0
transformers>=4.37.0,<5
